Fix various warnings in examples.
Uninitialized variables, unused args, and signedness
mismatches.nehe26 got a bit deeper of a cleanup in
order to address unchecked allocations.

+++ examples/dreamcast/2ndmix/2ndmix.c
@@ -731,7 +731,7 @@ void font_one_frame(void) {
 }
 
 void font_init(void) {
-    int x, y, c;
+    size_t x, y, c;
     uint8 pcxpall[768];
     volatile uint16 *vtex;
     uint16 val;
